+++ data/db/core/src/commonMain/kotlin/app/campfire/db/DatabaseFactory.kt
@@ -68,7 +68,6 @@ class DatabaseFactory(
       numInvalidAudioFilesAdapter = IntColumnAdapter,
       propertySizeAdapter = IntColumnAdapter,
       metadata_genresAdapter = StringListAdapter,
-      metadata_series_sequenceAdapter = IntColumnAdapter,
       metadata_seriesAdapter = SeriesSequenceListAdapter,
     )