+++ features/series/impl/src/commonMain/kotlin/app/campfire/series/paging/SeriesRemoteMediator.kt
@@ -123,7 +123,7 @@ class SeriesRemoteMediator(
               // Insert the series books
               series.books?.forEachIndexed { index, book ->
                 val libraryItem = book.asDbModel(user.serverUrl)
-                val media = book.media.asDbModel(book.id, fallbackSeriesSequence = index + 1)
+                val media = book.media.asDbModel(book.id, fallbackSeriesSequence = (index + 1).toDouble())
 
                 // If these items exist, lets not overwrite their metadata
                 db.libraryItemsQueries.insertOrIgnore(libraryItem)