+++ features/series/impl/src/commonMain/kotlin/app/campfire/series/store/SeriesSourceOfTruthFactory.kt
@@ -59,7 +59,7 @@ internal class SeriesSourceOfTruthFactory(
             // Insert the series books
             series.books?.forEachIndexed { index, book ->
               val libraryItem = book.asDbModel(userSession.serverUrl)
-              val media = book.media.asDbModel(book.id, fallbackSeriesSequence = index + 1)
+              val media = book.media.asDbModel(book.id, fallbackSeriesSequence = (index + 1).toDouble())
 
               // If these items exist, lets not overwrite their metadata
               db.libraryItemsQueries.insertOrIgnore(libraryItem)